FIX: Skip redirect_to_login_if_required if blocked
It does not make sense to redirect user to login if they are blocked.
This was causing issues for sites with login_required enabled because
they had to be logged in to see the error, but they were stuck in a
redirect loop because redirect_to_login_if_required was called over and
over by the initial controller and then by GeoblockingController too.